updating tyk example
