+++ examples/quick-start-tyk/.env.template
@@ -1,4 +1,7 @@
-TRACETEST_API_TOKEN=
-TRACETEST_AGENT_API_KEY=
+# Get the required information here: https://app.tracetest.io/retrieve-token
+
+TRACETEST_TOKEN="<YOUR_TRACETEST_TOKEN>"
+TRACETEST_ENVIRONMENT_ID="<YOUR_ENV_ID>"
+
 POKESHOP_DEMO_URL=http://tyk-gateway:8080
 TYK_AUTH_KEY=28d220fd77974a4facfb07dc1e49c2aa
