+++ examples/quick-start-tyk/otel-collector.yml
@@ -12,8 +12,8 @@ processors:
   batch:
 
 exporters:
-  jaeger:
-    endpoint: jaeger:14250
+  otlp/jaeger:
+    endpoint: jaeger:4317
     tls:
       insecure: true
   otlp:
@@ -26,7 +26,7 @@ service:
     traces:
       receivers: [otlp]
       processors: []
-      exporters: [jaeger]
+      exporters: [otlp/jaeger]
     traces/1:
       receivers: [otlp]
       processors: [batch]
