--- examples/quick-start-tyk/package.json
@@ -1,23 +1,0 @@
-{
-  "name": "quick-start-tyk",
-  "version": "1.0.0",
-  "description": "## Quick Start",
-  "main": "index.js",
-  "scripts": {
-    "pw:open": "playwright test --ui",
-    "test": "playwright test"
-  },
-  "keywords": [],
-  "author": "",
-  "license": "ISC",
-  "devDependencies": {
-    "@playwright/test": "^1.41.2",
-    "@tracetest/playwright": "^0.0.37",
-    "@types/node": "^20.11.20",
-    "@types/node-fetch": "^2.6.11",
-    "dotenv": "^16.4.5"
-  },
-  "dependencies": {
-    "node-fetch": "^2.7.0"
-  }
-}
