+++ examples/quick-start-tyk/resources/datastore.yaml
@@ -1,0 +1,12 @@
+type: DataStore
+spec:
+  id: current
+  name: jaeger
+  type: jaeger
+  default: true
+  jaeger:
+    endpoint: jaeger:16685
+    headers:
+      "": ""
+    tls:
+      insecure: true
